impl ability to view xdsm from a recording file
