+++ openmdao/devtools/xdsm_viewer/xdsm_writer.py
@@ -1561,89 +1561,3 @@ def _multiline_block(*texts, **kwargs):
     template = '$\\begin{{array}}{{{pos}}} {text} \\end{{array}}$'
     new_line = ' \\\\ '
     return template.format(text=new_line.join(texts), pos='c'*len(texts))
-
-
-##### openmdao command line setup
-
-
-def _xdsm_setup_parser(parser):
-    """
-    Set up the openmdao subparser for the 'openmdao xdsm' command.
-
-    Parameters
-    ----------
-    parser : argparse subparser
-        The parser we're adding options to.
-    """
-    parser.add_argument('file', nargs=1, help='Python file containing the model.')
-    parser.add_argument('-o', '--outfile', default='xdsm_out', action='store', dest='outfile',
-                        help='XDSM output file. (use pathname without extension)')
-    parser.add_argument('-f', '--format', default='html', action='store', dest='format',
-                        choices=['html', 'pdf', 'tex'], help='format of XSDM output.')
-    parser.add_argument('-m', '--model_path', action='store', dest='model_path',
-                        help='Path to system to transcribe to XDSM.')
-    parser.add_argument('-r', '--recurse', action='store_true', dest='recurse',
-                        help="Don't treat the top level of each name as the source/target component.")
-    parser.add_argument('--no_browser', action='store_true', dest='no_browser',
-                        help="Don't display in a browser.")
-    parser.add_argument('--no_parallel', action='store_true', dest='no_parallel',
-                        help="don't show stacked parallel blocks. Only active for 'pdf' and 'tex' "
-                             "formats.")
-    parser.add_argument('--no_ext', action='store_true', dest='no_extern_outputs',
-                        help="Don't show externally connected outputs.")
-    parser.add_argument('-s', '--include_solver', action='store_true', dest='include_solver',
-                        help="Include the problem model's solver in the XDSM.")
-    parser.add_argument('--no_process_conns', action='store_true', dest='no_process_conns',
-                        help="Don't add process connections (thin black lines).")
-    parser.add_argument('--box_stacking', action='store', default=_DEFAULT_BOX_STACKING,
-                        choices=['max_chars', 'vertical', 'horizontal', 'cut_chars', 'empty'],
-                        dest='box_stacking', help='Controls the appearance of boxes.')
-    parser.add_argument('--box_width', action='store', default=_DEFAULT_BOX_WIDTH,
-                        dest='box_width', type=int, help='Controls the width of boxes.')
-    parser.add_argument('--box_lines', action='store', default=_MAX_BOX_LINES,
-                        dest='box_lines', type=int,
-                        help='Limits number of vertical lines in box if box_stacking is vertical.')
-    parser.add_argument('--numbered_comps', action='store_true', dest='numbered_comps',
-                        help="Display components with numbers.  Only active for 'pdf' and 'tex' "
-                        "formats.")
-    parser.add_argument('--number_alignment', action='store', dest='number_alignment',
-                        choices=['horizontal', 'vertical'], default='horizontal',
-                        help='Positions the number either above or in front of the component label '
-                        'if numbered_comps is true.')
-    parser.add_argument('--output_side', action='store', dest='output_side', default=_DEFAULT_OUTPUT_SIDE,
-                        help='Position of the outputs on the diagram. Left or right, or a '
-                             'dictionary with component types as keys. Component type key can be '
-                             '"optimization", "doe" or "default".')
-
-
-def _xdsm_cmd(options):
-    """
-    Return the post_setup hook function for 'openmdao xdsm'.
-
-    Parameters
-    ----------
-    options : argparse Namespace
-        Command line options.
-
-    Returns
-    -------
-    function
-        The post-setup hook function.
-    """
-    def _xdsm(prob):
-        kwargs = {}
-        for name in ['box_stacking', 'box_width', 'box_lines', 'numbered_comps', 'number_alignment']:
-            val = getattr(options, name)
-            if val is not None:
-                kwargs[name] = val
-
-        write_xdsm(prob, filename=options.outfile, model_path=options.model_path,
-                   recurse=options.recurse,
-                   include_external_outputs=not options.no_extern_outputs,
-                   out_format=options.format,
-                   include_solver=options.include_solver, subs=_CHAR_SUBS,
-                   show_browser=not options.no_browser, show_parallel=not options.no_parallel,
-                   add_process_conns=not options.no_process_conns, output_side=options.output_side,
-                   **kwargs)
-        exit()
-    return _xdsm
